+++ application/views/template.php
@@ -23,7 +23,15 @@
                     <li><a href="">Equipments</a>
                         <ul class="dropdown-menu" role="menu">
                             <li><a role="menuitem" href="/presetscontroller">Presets</a></li>
-                            <li><a role="menuitem" href="/equipmentscontroller">Equipments</a></li>
+                            <li><a role="menuitem" href="/equipmentscontroller">All Equipments</a></li>
+                            <li><a role="menuitem" href="/category/amulet">Amulet</a></li>
+                            <li><a role="menuitem" href="/category/belt">Belt</a></li>
+                            <li><a role="menuitem" href="/category/boot">Boots</a></li>
+                            <li><a role="menuitem" href="/category/glove">Gloves</a></li>
+                            <li><a role="menuitem" href="/category/helmet">Helmet</a></li>
+                            <li><a role="menuitem" href="/category/chest">Chest</a></li>
+                            <li><a role="menuitem" href="/category/ring">Ring</a></li>
+                            <li><a role="menuitem" href="/category/weapon">Weapon</a></li>
                         </ul>
                     </li>
                     <li><a href="/history">History</a></li>